app/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.7.2"
pydantic-settings = "^2.2.1"
selenium = "^4.21.0"
boto3 = "^1.34.121"
fastapi = {extras = ["all"], version = "^0.111.0"}
uvicorn = {extras = ["standard"], version = "^0.30.1"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
jpholiday = "^0.1.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
